pyproject.toml
[project]
name = "mobile-de-parsing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9, <3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "requests>=2.32.4",
    "selenium>=4.34.2",
]

[tool.uv.workspace]
members = [
    "requests",
]
